Move all maps config to helper class
Switch to helper class for following maps:
- logbook
- moorages
- default map (dashboard)
