+++ src/pages/logs-book/DetailsMap.vue
@@ -108,6 +108,7 @@
   import 'leaflet-sidebar-v2/css/leaflet-sidebar.min.css'
   import * as L from 'leaflet'
   import 'leaflet-sidebar-v2'
+  import { defaultBaseMapType, baseMaps, overlayMaps } from '../../utils/leafletHelpers.js'
 
   import { computed, ref, reactive, onMounted } from 'vue'
   import { useRoute } from 'vue-router'
@@ -254,66 +255,12 @@
     }).setView([centerLat, centerLng], 17)
     // Add new Zoom control
     L.control.zoom({ position: 'bottomright' }).addTo(map.value)
-    // OSM
-    const osm = L.tileLayer('http://{s}.tile.osm.org/{z}/{x}/{y}.png', {
-      attribution: '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors',
-      maxZoom: 18,
-    })
-    // OpenSeaMap
-    const openseamap = L.tileLayer('https://tiles.openseamap.org/seamark/{z}/{x}/{y}.png', {
-      attribution: '&copy; <a href="http://www.openseamap.org">OpenSeaMap</a> contributors',
-      maxZoom: 18,
-    })
-    // Satellite
-    const sat = L.tileLayer(
-      'https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}',
-      {
-        attribution:
-          'Tiles &copy; Esri &mdash; Source: Esri, i-cubed, USDA, USGS, AEX, GeoEye, Getmapping, Aerogrid, IGN, IGP, UPR-EGP, and the GIS User Community',
-        maxZoom: 17,
-      },
-    )
-    // NOAA
-    const noaa = L.tileLayer('https://tileservice.charts.noaa.gov/tiles/50000_1/{z}/{x}/{y}.png', {
-      attribution: 'NOAA',
-      maxZoom: 18,
-    })
-    const cartodb = L.tileLayer('https://{s}.basemaps.cartocdn.com/light_all/{z}/{x}/{y}.png', {
-      attribution:
-        '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, &copy; <a href="https://carto.com/attribution">CARTO</a>',
-      maxZoom: 18,
-    })
-    // https://emodnet.ec.europa.eu
-    const bathymetryLayer = L.tileLayer.wms('http://ows.emodnet-bathymetry.eu/wms', {
-      layers: 'emodnet:mean_atlas_land',
-      format: 'image/png',
-      transparent: true,
-      attribution: 'EMODnet Bathymetry',
-      opacity: 0.8,
-    })
-    const coastlinesLayer = L.tileLayer.wms('http://ows.emodnet-bathymetry.eu/wms', {
-      layers: 'coastlines',
-      format: 'image/png',
-      transparent: true,
-      attribution: 'EMODnet Bathymetry',
-      opacity: 0.8,
-    })
-    const bathymetryGroupLayer = L.layerGroup([bathymetryLayer, coastlinesLayer])
-    //bathymetryGroupLayer.addTo(map)
 
-    const baseMaps = {
-      OpenStreetMap: osm,
-      Satellite: sat,
-      NOAA: noaa,
-      Carto: cartodb,
-      'EMODnet Bathymetry': bathymetryGroupLayer,
-    }
-    const overlays = {
-      OpenSeaMap: openseamap,
-    }
-    L.control.layers(baseMaps, overlays).addTo(map.value)
-    baseMaps['OpenStreetMap'].addTo(map.value)
-    openseamap.addTo(map.value)
+    const bMaps = baseMaps()
+    const oMaps = overlayMaps()
+    bMaps[defaultBaseMapType()].addTo(map.value)
+
+    L.control.layers(bMaps, oMaps).addTo(this.map)
 
     const sailConfigIconImg = function (feature) {
       if (
